Fixed calculation of lines per second

+++ src/main/java/com/xceptance/xlt/report/DataRecordReader.java
@@ -6,7 +6,7 @@ import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ConcurrentSkipListMap;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -50,7 +50,7 @@ class DataRecordReader implements Runnable
     /**
      * The global line counter.
      */
-    private final AtomicInteger totalLineCounter;
+    private final AtomicLong totalLineCounter;
 
     /**
      * The instance number of the test user.
@@ -79,7 +79,7 @@ class DataRecordReader implements Runnable
      *            the dispatcher that coordinates result processing
      */
     public DataRecordReader(final FileObject directory, final String agentName, final String testCaseName, final String userNumber,
-                            final AtomicInteger totalLineCounter, final Dispatcher dispatcher)
+                            final AtomicLong totalLineCounter, final Dispatcher dispatcher)
     {
         this.directory = directory;
         this.agentName = agentName;
